Part-2-Interactive QA Bot Interface/requirements.txt
streamlit
PyPDF2
cohere
pinecone-client
sentence-transformers